refactor: rename package to correct vocab
