+++ internal/handlers/tracks/handler.go
@@ -5,13 +5,13 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/jetaimejeteveux/music-catalog/internal/middleware"
 	spotifyModel "github.com/jetaimejeteveux/music-catalog/internal/models/spotify"
-	"github.com/jetaimejeteveux/music-catalog/internal/models/trackactivites"
+	"github.com/jetaimejeteveux/music-catalog/internal/models/trackactivities"
 )
 
 //go:generate mockgen -source=handler.go -destination=handler_mock.go -package=tracks
 type service interface {
 	Search(ctx context.Context, query string, pageSize, pageIndex int, userId uint) (*spotifyModel.SearchResponse, error)
-	UpsertTrackActivities(ctx context.Context, userId uint, request trackactivites.TrackActivityReqest) error
+	UpsertTrackActivities(ctx context.Context, userId uint, request trackactivities.TrackActivityReqest) error
 }
 
 type Handler struct {
