+++ internal/repository/trackactivities/create.go
@@ -2,9 +2,9 @@ package trackactivities
 
 import (
 	"context"
-	"github.com/jetaimejeteveux/music-catalog/internal/models/trackactivites"
+	"github.com/jetaimejeteveux/music-catalog/internal/models/trackactivities"
 )
 
-func (r *repository) Create(ctx context.Context, model trackactivites.TrackActivity) error {
+func (r *repository) Create(ctx context.Context, model trackactivities.TrackActivity) error {
 	return r.db.Create(&model).Error
 }
