+++ internal/repository/trackactivities/getBulkSpotifyIDs.go
@@ -2,16 +2,16 @@ package trackactivities
 
 import (
 	"context"
-	"github.com/jetaimejeteveux/music-catalog/internal/models/trackactivites"
+	"github.com/jetaimejeteveux/music-catalog/internal/models/trackactivities"
 )
 
-func (r *repository) GetBulkSpotifyIDs(ctx context.Context, userId uint, spotifyIds []string) (map[string]trackactivites.TrackActivity, error) {
-	var models []trackactivites.TrackActivity
+func (r *repository) GetBulkSpotifyIDs(ctx context.Context, userId uint, spotifyIds []string) (map[string]trackactivities.TrackActivity, error) {
+	var models []trackactivities.TrackActivity
 	res := r.db.Where("user_id = ?", userId).Where("spotify_id IN (?)", spotifyIds).Find(&models)
 	if err := res.Error; err != nil {
 		return nil, err
 	}
-	result := make(map[string]trackactivites.TrackActivity, 0)
+	result := make(map[string]trackactivities.TrackActivity, 0)
 	for _, model := range models {
 		result[model.SpotifyId] = model
 	}
