+++ internal/service/tracks/service.go
@@ -2,7 +2,7 @@ package tracks
 
 import (
 	"context"
-	"github.com/jetaimejeteveux/music-catalog/internal/models/trackactivites"
+	"github.com/jetaimejeteveux/music-catalog/internal/models/trackactivities"
 	"github.com/jetaimejeteveux/music-catalog/internal/repository/spotify"
 )
 
@@ -12,10 +12,10 @@ type SpotifyOutbond interface {
 }
 
 type trackActivitiesRepository interface {
-	Create(ctx context.Context, model trackactivites.TrackActivity) error
-	Update(ctx context.Context, model trackactivites.TrackActivity) error
-	Get(ctx context.Context, userId uint, spotifyId string) (*trackactivites.TrackActivity, error)
-	GetBulkSpotifyIDs(ctx context.Context, userId uint, spotifyIds []string) (map[string]trackactivites.TrackActivity, error)
+	Create(ctx context.Context, model trackactivities.TrackActivity) error
+	Update(ctx context.Context, model trackactivities.TrackActivity) error
+	Get(ctx context.Context, userId uint, spotifyId string) (*trackactivities.TrackActivity, error)
+	GetBulkSpotifyIDs(ctx context.Context, userId uint, spotifyIds []string) (map[string]trackactivities.TrackActivity, error)
 }
 type service struct {
 	spotifyOutbond      SpotifyOutbond
